// src/interfaces/tags.interface.ts
export interface ITag{
    name: string;
    className: string;
    icon: any;
}

export interface ITags{
    angular: ITag;
    laravel: ITag;
    typescript: ITag;
    php: ITag;
    mysql: ITag;
    tailwind: ITag;
    bootstrap: ITag;
    git: ITag;
    github: ITag;
    react: ITag;
    astro: ITag;
    svelte: ITag;
    javascript: ITag;
    html: ITag;
    css: ITag;
}